+++ addons/survey/__manifest__.py
@@ -18,19 +18,27 @@ sent mails with personal token for the invitation of the survey.
     'website': 'https://www.odoo.com/page/survey',
     'depends': ['http_routing', 'mail'],
     'data': [
+        'data/mail_template_data.xml',
+        'data/survey_data.xml',
+        'data/ir_actions_data.xml',
         'security/survey_security.xml',
         'security/ir.model.access.csv',
+        'views/assets.xml',
+        'views/survey_menus.xml',
+        'views/survey_stage_views.xml',
         'views/survey_survey_views.xml',
+        'views/survey_user_views.xml',
+        'views/survey_question_views.xml',
+        'views/survey_page_views.xml',
         'views/survey_templates.xml',
-        'views/survey_result.xml',
-        'data/mail_template_data.xml',
         'wizard/survey_mail_compose_message_views.xml',
-        'data/survey_data.xml',
     ],
-    'demo': ['data/survey_demo_user.xml',
-             'data/survey_demo_feedback.xml',
-             'data/survey.user_input.csv',
-             'data/survey.user_input_line.csv'],
+    'demo': [
+        'data/survey_demo_user.xml',
+        'data/survey_demo_feedback.xml',
+        'data/survey.user_input.csv',
+        'data/survey.user_input_line.csv'
+    ],
     'installable': True,
     'auto_install': False,
     'application': True,
