+++ addons/survey/models/__init__.py
@@ -1,5 +1,9 @@
 # -*- encoding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import survey_survey
 from . import ir_autovacuum
+from . import survey_survey
+from . import survey_page
+from . import survey_question
+from . import survey_stage
+from . import survey_user
