+++ addons/survey/views/survey_menus.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+<data>
+    <!-- Main menu -->
+    <menuitem name="Surveys" id="menu_surveys" sequence="70" groups="group_survey_user"  web_icon="survey,static/description/icon.png"/>
+
+    <!-- Parent menus -->
+    <menuitem name="Configuration" id="menu_surveys_configuration" parent="menu_surveys" groups="base.group_no_one" sequence="100"/>
+</data>
+</odoo>
