+++ addons/survey/views/survey_page_views.xml
@@ -1,0 +1,71 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+<data>
+    <!-- PAGES -->
+    <record model="ir.ui.view" id="survey_page_form">
+        <field name="name">Form view for survey page</field>
+        <field name="model">survey.page</field>
+        <field name="arch" type="xml">
+            <form string="Survey Page" create="false">
+                <sheet>
+                    <div class="oe_title" style="width: 100%;">
+                        <label for="title" class="oe_edit_only"/>
+                        <h1><field name="title" placeholder="Page Title"/></h1>
+                    </div>
+                    <field name="question_ids" nolabel="1" mode="tree" context="{'default_page_id': active_id}">
+                        <tree>
+                            <field name="sequence" widget="handle"/>
+                            <field name="question"/>
+                            <field name="question_type"/>
+                            <field name="constr_mandatory"/>
+                        </tree>
+                    </field>
+                </sheet>
+            </form>
+        </field>
+    </record>
+    <record model="ir.ui.view" id="survey_page_tree">
+        <field name="name">Tree view for survey page</field>
+        <field name="model">survey.page</field>
+        <field name="arch" type="xml">
+            <tree string="Survey Page" create="false">
+                <field name="sequence" widget="handle"/>
+                <field name="title"/>
+                <field name="survey_id"/>
+                <field name="question_ids" string="#Questions"/>
+            </tree>
+        </field>
+    </record>
+    <record model="ir.ui.view" id="survey_page_search" >
+        <field name="name">Search view for survey page</field>
+        <field name="model">survey.page</field>
+        <field name="arch" type="xml">
+            <search string="Search Page">
+                <field name="title" string="Page"/>
+                <field name="survey_id" string="Survey"/>
+                <group expand="0" string="Group By">
+                    <filter name="group_by_survey" string="Survey" domain="[]"  context="{'group_by':'survey_id'}"/>
+                </group>
+            </search>
+        </field>
+    </record>
+
+    <record model="ir.actions.act_window" id="action_survey_page_form">
+        <field name="name">Pages</field>
+        <field name="res_model">survey.page</field>
+        <field name="view_type">form</field>
+        <field name="view_mode">tree,form</field>
+        <field name="search_view_id" ref="survey_page_search"/>
+        <field name="context">{'search_default_group_by_survey': True}</field>
+        <field name="help" type="html">
+          <p class="o_view_nocontent_empty_folder">
+            No pages found
+          </p>
+        </field>
+    </record>
+
+    <act_window context="{'search_default_page_id': active_id, 'default_page_id': active_id}" id="act_survey_page_question" name="Questions" res_model="survey.question" src_model="survey.page"/>
+
+    <menuitem name="Pages" id="menu_survey_page_form1" action="action_survey_page_form" parent="menu_surveys_configuration" sequence="1"/>
+</data>
+</odoo>
