+++ addons/survey/views/survey_stage_views.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+<data>
+    <record model="ir.ui.view" id="survey_stage_form">
+        <field name="name">Form view for survey stage</field>
+        <field name="model">survey.stage</field>
+        <field name="arch" type="xml">
+            <form string="Stage">
+                <group col='4' colspan='4'>
+                <field name="name" />
+                <field name="closed" />
+                <field name="sequence" />
+                <field name="fold" />
+                </group>
+            </form>
+        </field>
+    </record>
+</data>
+</odoo>
